requirements.txt
telethon>=1.36.0
mcp>=1.0.0
qrcode>=7.0
python-dotenv>=1.0.0
